refactor(domain/tasks_lis): rename exceptions file
