+++ todoapp/domain/tasks_list/entities.py
@@ -6,7 +6,7 @@ from uuid6 import uuid7
 
 from todoapp.domain.common.constants import Operation
 from todoapp.domain.common.entities import BaseEntity
-from todoapp.domain.tasks_list.exception import TaskListVisibilityNotModified, SharingRuleNotExistsError
+from todoapp.domain.tasks_list.exceptions import TaskListVisibilityNotModified, SharingRuleNotExistsError
 from todoapp.domain.tasks_list.value_objects import ListId, Sharing, SharingRule
 from todoapp.domain.user.entities import UserId
